Move construct_data to another file and rename api.py to evaluator.py
