+++ src/gluonts/ev/__init__.py
@@ -28,16 +28,15 @@ from .metrics import (
     WeightedSumQuantileLoss,
 )
 from .aggregations import Aggregation, Sum, Mean
-from .api import (
+from .evaluator import (
     Metric,
     MetricEvaluator,
     MultiMetricEvaluator,
     DerivedMetricEvaluator,
     StandardMetricEvaluator,
-    construct_data,
 )
 from .stats import seasonal_error
-
+from .data_preparation import construct_data
 
 __all__ = [
     "mean_absolute_label",
